msi file
In order to package a Keyman keyboard the system asks for a msi file. I searched high and low for it until I found the file "keymandesktop100.msi", but when i tried to save it, it could not be found anywhere in the computer. Is this the right file and if so what do I do with it?